requirements.txt
torch
torchaudio
openai-whisper
pyaudio
numpy